chore: extract common logic to a custom hook
