+++ plugins/main/public/components/common/wazuh-discover/table.tsx
@@ -1,10 +1,4 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import { IntlProvider } from 'react-intl';
 import {
   EuiDataGrid,
@@ -44,10 +38,7 @@ import { LoadingSearchbarProgress } from '../loading-searchbar-progress/loading-
 // common components/hooks
 import { IndexPattern } from '../../../../../../../src/plugins/data/public';
 import { wzDiscoverRenderColumns } from './render-columns';
-import {
-  applyHitSourceUpdate,
-  applyResultsSourceUpdate,
-} from './apply-hit-source-update';
+import { useDocumentMutationSync } from './use-document-mutation-sync';
 import {
   DocumentViewTableAndJson,
   DocumentViewTableAndJsonPropsAdditionalTabs,
@@ -316,11 +307,11 @@ export const useTableDataGridFetch = ({
 
   const [results, setResults] = useState<SearchResponse>({} as SearchResponse);
   const [inspectedHit, setInspectedHit] = useState<any>(undefined);
-  const [caseRefreshKey, setCaseRefreshKey] = useState(0);
-
-  // Keep the latest inspected hit reachable from the stable mutation handler.
-  const inspectedHitRef = useRef(inspectedHit);
-  inspectedHitRef.current = inspectedHit;
+  const { onDocumentMutated, caseRefreshKey } = useDocumentMutationSync({
+    inspectedHit,
+    setInspectedHit,
+    setResults,
+  });
 
   const onClickInspectDoc = useMemo(
     () => (index: number) => {
@@ -328,28 +319,6 @@ export const useTableDataGridFetch = ({
       setInspectedHit(rowClicked);
     },
     [results],
-  );
-
-  // Called when the open document is mutated from its flyout (e.g. a case
-  // create/edit/clean). Patches both the flyout and its row in the grid with the
-  // returned payload so they update instantly and stay consistent, and bumps a key
-  // that refreshes the dashboard charts (see InventoryDashboard). The grid is updated
-  // optimistically here instead of refetching, since a _search would race the index
-  // refresh.
-  const onDocumentMutated = useCallback(
-    (sourceUpdate?: Record<string, unknown>) => {
-      if (sourceUpdate) {
-        const documentId = inspectedHitRef.current?._id;
-        setInspectedHit(current => applyHitSourceUpdate(current, sourceUpdate));
-        if (documentId) {
-          setResults(current =>
-            applyResultsSourceUpdate(current, documentId, sourceUpdate),
-          );
-        }
-      }
-      setCaseRefreshKey(key => key + 1);
-    },
-    [],
   );
 
   const DocViewInspectButton = ({
