+++ plugins/main/public/components/common/wazuh-discover/wz-discover.tsx
@@ -1,10 +1,4 @@
-import React, {
-  useState,
-  useMemo,
-  useEffect,
-  useCallback,
-  useRef,
-} from 'react';
+import React, { useState, useMemo, useEffect } from 'react';
 import {
   EuiDataGrid,
   EuiPageTemplate,
@@ -57,10 +51,7 @@ import { wzDiscoverRenderColumns } from './render-columns';
 import { WzSearchBar } from '../search-bar';
 import { transformDateRange } from '../search-bar/search-bar-service';
 import DocDetailsHeader from './components/doc-details-header';
-import {
-  applyHitSourceUpdate,
-  applyResultsSourceUpdate,
-} from './apply-hit-source-update';
+import { useDocumentMutationSync } from './use-document-mutation-sync';
 import { tDataGridColumn } from '../data-grid/types';
 import { SampleDataWarning } from '../../visualize/components';
 
@@ -92,32 +83,8 @@ const WazuhDiscoverComponent = (props: WazuhDiscoverProps) => {
 
   const [results, setResults] = useState<SearchResponse>({} as SearchResponse);
   const [inspectedHit, setInspectedHit] = useState<any>(undefined);
-  const [caseRefreshKey, setCaseRefreshKey] = useState(0);
-
-  // Keep the latest inspected hit reachable from the stable mutation handler.
-  const inspectedHitRef = useRef(inspectedHit);
-  inspectedHitRef.current = inspectedHit;
-
-  // Called when the open document is mutated from its flyout.
-  // Patches both the flyout and its row in the grid with the
-  // returned payload so they update instantly and stay consistent, and bumps a key
-  // that refreshes the dashboard charts. The grid is updated optimistically here
-  // instead of refetching, since a _search would race the index refresh.
-  const handleDocumentMutated = useCallback(
-    (sourceUpdate?: Record<string, unknown>) => {
-      if (sourceUpdate) {
-        const documentId = inspectedHitRef.current?._id;
-        setInspectedHit(current => applyHitSourceUpdate(current, sourceUpdate));
-        if (documentId) {
-          setResults(current =>
-            applyResultsSourceUpdate(current, documentId, sourceUpdate),
-          );
-        }
-      }
-      setCaseRefreshKey(key => key + 1);
-    },
-    [],
-  );
+  const { onDocumentMutated: handleDocumentMutated, caseRefreshKey } =
+    useDocumentMutationSync({ inspectedHit, setInspectedHit, setResults });
   const [indexPattern, setIndexPattern] = useState<IndexPattern | undefined>(
     undefined,
   );
